pyproject.toml
[tool.poetry]
name = "github-vuln-scraper"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
PyGithub = "^1.55"
GitPython = "^3.1.27"
bandit = "^1.7.4"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
